+++ share/objects/specials/blood-stone/02-ancient-blood-stone.lua
@@ -13,12 +13,12 @@ do
 			.. ' <A019,DataA%d>%% to do <A019,DataB%d,.>x damage.|n'
 
 			.. '-'
-			.. ' 24%% chance to cast a <A07A,DataA%d> damage Blood'
+			.. ' 24%% chance to cast a <A07A,DataA%d> damage Ancient Blood'
 			.. ' Lightning spell, costing <A07A,Cost%d> mana.|n'
 
 			.. '-'
-			.. ' +12%% chance to cast Blood Lightning for each additional'
-			.. ' Blood Stone or Ancient Blood Stone.|n'
+			.. ' +12%% chance to cast Ancient Blood Lightning for each'
+			.. ' additional Blood Stone or Ancient Blood Stone.|n'
 
 		for level = 1, 11 do
 			ability.aub1.values [level] = string.format (